Fix operate and operateClick bugs
When operate was called, it did not reset the values of the equation.
Changed this bug so it now correctly resets values.

In the event where the user selects an operator when both numbers are
assigned values, operateClick did not use the result of the equation as
a number in the next equation. Fixed by changing operateClick to assign
the result of the previous equation to firstNum.